# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/fossil": {
            "get": {
                "description": "Возвращает список всех не удаленных ископаемых",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Останки"
                ],
                "summary": "Получение списка ископаемых",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Название вида",
                        "name": "searchSpecies",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Начало даты формирования",
                        "name": "startFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Конец даты формирования",
                        "name": "endFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Статус ископаемого",
                        "name": "fossilStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список ископаемых",
                        "schema": {
                            "$ref": "#/definitions/model.FossilRequest"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FossilRequest"
                        }
                    }
                }
            }
        },
        "/fossil/{id}": {
            "get": {
                "description": "Возвращает информацию об останке по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Останки"
                ],
                "summary": "Получение ископаемого по идентификатору",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор ископаемого",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об останке",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор ископаемого",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    }
                }
            }
        },
        "/fossil/{id}/delete": {
            "delete": {
                "description": "Удаляет ископаемое по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Останки"
                ],
                "summary": "Удаление ископаемого",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор ископаемого",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Название вида",
                        "name": "searchSpecies",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Начало даты формирования",
                        "name": "startFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Конец даты формирования",
                        "name": "endFormationDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Статус ископаемого",
                        "name": "fossilStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список периодов",
                        "schema": {
                            "$ref": "#/definitions/model.FossilRequest"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор ископаемого",
                        "schema": {
                            "$ref": "#/definitions/model.FossilRequest"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FossilRequest"
                        }
                    }
                }
            }
        },
        "/fossil/{id}/status": {
            "put": {
                "description": "Обновляет статус ископаемого для модератора по идентификатору ископаемого",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Останки"
                ],
                "summary": "Обновление статуса ископаемого для модератора",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор ископаемого",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус ископаемого",
                        "name": "fossilStatus",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FossilUpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об останке",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор ископаемого или ошибка чтения JSON объекта",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    }
                }
            }
        },
        "/fossil/{id}/update": {
            "put": {
                "description": "Обновляет номер рейса для ископаемого по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Останки"
                ],
                "summary": "Обновление номера рейса ископаемого",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор ископаемого",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый вид ископаемого",
                        "name": "species",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FossilUpdateSpeciesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об останке",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор ископаемого или ошибка чтения JSON объекта",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    }
                }
            }
        },
        "/fossil/{id}/user": {
            "put": {
                "description": "Обновляет статус ископаемого для пользователя по идентификатору ископаемого",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Останки"
                ],
                "summary": "Обновление статуса ископаемого для пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор ископаемого",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об останке",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    },
                    "400": {
                        "description": "Недопустимый идентификатор ископаемого",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.FossilGetResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.FossilGetResponse": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id_fossil": {
                    "type": "integer"
                },
                "periods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Period"
                    }
                },
                "species": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.FossilRequest": {
            "type": "object",
            "properties": {
                "completion_date": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "formation_date": {
                    "type": "string"
                },
                "id_fossil": {
                    "type": "integer"
                },
                "species": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.FossilUpdateSpeciesRequest": {
            "type": "object",
            "properties": {
                "species": {
                    "type": "string"
                }
            }
        },
        "model.FossilUpdateStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "model.Period": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id_period": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}